Title: 최소 통과 시간, Time: 7ms, Memory: 0MB, Status: Passed - Codetree

+++ 260308/최소 통과 시간/minimum-transit-time.cpp
@@ -33,7 +33,7 @@ int main() {
     }
 
     long long left = 1;
-    long long right = n * maxTime;
+    long long right = (long long)n * maxTime;
 
     while (left <= right)
     {
